see description for further plans
Currently, we can't get the time that we released the mouse on. But we could potentially click the mouse again right away. Use a global bool to know if this second press was the release label or the press label and than if we have got both, we can continue with what we were doing